backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
pillow>=11.0.0

# tests
httpx>=0.28.0
pytest>=8.3.0
